Methods/requirements.txt

python-rake
pandas
